feat: 트래픽 제어 zone 현황 조회 service 기능 개발 추가

+++ ktca/src/main/kotlin/com/kona/ktca/v1/domain/model/TrafficZoneWaiting.kt
@@ -1,7 +1,7 @@
 package com.kona.ktca.v1.domain.model
 
 data class TrafficZoneWaiting(
-    val waitingCount: Int = 0,
-    val entryCount: Int = 0,
-    val estimatedClearTime: Int = 0,
+    val waitingCount: Long = 0,
+    val entryCount: Long = 0,
+    val estimatedClearTime: Long = 0,
 )
